CDK/requirements.txt
aws-cdk.core==1.62.0
aws-cdk.aws-cognito==1.62.0
aws-cdk.aws-apigateway==1.62.0
aws-cdk.aws-lambda==1.62.0
aws-cdk.aws-iam==1.62.0
aws-cdk.aws-cognito==1.62.0
aws-cdk.aws-apigateway==1.62.0